refactor(cta): extrair CtaBody — CallToAction reduzido a orquestrador de layout
